+++ backend/src/devices/artnet/index.ts
@@ -1,7 +1,11 @@
 import { Artnet } from 'artnet'
 import { logger } from '@vlight/utils'
 
-import { artnetHost, enableArtNetDevices } from '../../services/config'
+import {
+  artnetHost,
+  artnetUniverse,
+  enableArtNetDevices,
+} from '../../services/config'
 import { getDmxUniverse } from '../../services/universe'
 import { howLong } from '../../util/time'
 import { deviceRegistry } from '../registry'
@@ -13,7 +17,7 @@ function broadcastUniverseChannel(channel: number, value: number) {
     return
   }
   logger.trace('set ArtNet channel', channel, value)
-  server.set(channel, value)
+  server.set(artnetUniverse, channel, value)
 }
 
 export async function init(): Promise<void> {
@@ -27,7 +31,7 @@ export async function init(): Promise<void> {
   const artnet = require('artnet') // eslint-disable-line
 
   server = artnet({ host: artnetHost })
-  server.set(Array.from(getDmxUniverse()))
+  server.set(artnetUniverse, 1, Array.from(getDmxUniverse()))
 
   deviceRegistry.register({ broadcastUniverseChannel })