+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## March 2024
 
 - Added real-world memory preview during editing
+- Added `artnetUniverse` config
 - Fixed crash when deleting the last memory scene state
 
 ## February 2024